\name{l_genes}
\alias{l_genes}
\docType{data}
\title{
l_genes
}
\description{
Character vector containing the AGI gene codes of the genes used to reconstruct the network in Lavenus et al. 2015, Plant Cell.
}
\usage{data("l_genes")}

\examples{
# Load the vector
data(l_genes)

# Have a look at it
l_genes
}
\keyword{datasets}
