\name{l_names}
\alias{l_names}
\docType{data}
\title{
l_names
}
\description{
Character vector containing the 'everyday names' of the genes used to reconstruct the network in Lavenus et al. 2015, Plant Cell.
}

\usage{data("l_names")}

\examples{
# Load the vector
data(l_names)

# Have a look at it
l_names
}
\keyword{datasets}
