\name{times}
\alias{times}
\docType{data}
\title{
The \code{times} vector to use with the lateral root dataset
}
\description{
Contains the times (in hours) at which the samples were collected to generate the Lateral Root transcriptomic dataset (\code{data(LR_dataset)})
}
\usage{data("times")}

\examples{
# Load the vector associated with the LR dataset
data(times)

# Have a look at it
times
}
\keyword{datasets}
